refactor(core): 复用 validateWorkingDir 校验建群绑定目录并修复裸 ~ 展开
- 新增 src/core/working-dir.ts 叶子模块，承载 expandHome + validateWorkingDir，
  让 CLI 入口可直接引用而不牵入 daemon 依赖图（worker-pool/PTY）
- create-group CLI、dashboard IPC（含两处既有重复）统一改用 validateWorkingDir，
  删除手写的 trim/expand/resolve/exists/isDirectory 校验块
- 修复 cli.ts create-group 对裸 `~` 展开失败的 bug（原手写逻辑只处理 `~/`）
- session-manager / command-handler 改为 re-export，既有调用方不变

(review follow-up for PR #24)
